+++ server/src/auth/auth.service.ts
@@ -1,16 +1,12 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model, Types } from 'mongoose';
+import { Model } from 'mongoose';
 import { UserConflictException } from 'src/exceptions/conflict.exception';
 import { User, UserDocument } from 'src/user/schemas/user.schema';
 import { JwtService } from '@nestjs/jwt';
 import { LoginFailException } from 'src/exceptions/login-fail.exception';
 import { InvalidRefreshTokenException } from 'src/exceptions/invalid-refresh-token.exception';
 import { UserInfoDto } from 'src/user/dto/user-info.dto';
-import { createPresignedUrl } from 'src/ncpAPI/presignedURL';
-import * as _ from 'lodash';
-import { listObjects } from 'src/ncpAPI/listObjects';
-import { xml2js } from 'xml-js';
 import { SignupRequestDto } from './dto/signup-request.dto';
 import { JwtDto } from './dto/jwt.dto';
 import { SignupResponseDto } from './dto/signup-response.dto';
@@ -96,71 +92,5 @@ export class AuthService {
     return this.getLoginInfo(user).then(
       (loginInfo) => new SigninResponseDto(loginInfo),
     );
-  }
-
-  async getAdvertisementPresignedUrl(adName: string) {
-    if (adName)
-      return createPresignedUrl(
-        process.env.ADVERTISEMENT_BUCKET,
-        adName,
-        'GET',
-      );
-
-    const xmlData = await listObjects(process.env.ADVERTISEMENT_BUCKET);
-    const jsonData: any = xml2js(xmlData, { compact: true });
-
-    const adList = _.map(jsonData.ListBucketResult.Contents, 'Key._text');
-    const advertisements = await Promise.all(
-      adList.map(async (advertisement: string) => {
-        return createPresignedUrl(
-          process.env.ADVERTISEMENT_BUCKET,
-          advertisement,
-          'GET',
-        );
-      }),
-    );
-    return { advertisements };
-  }
-
-  async putProfilePresignedUrl({ uuid, profileExtension }) {
-    const objectName = `${uuid}.${profileExtension}`;
-    const presignedUrl = (
-      await createPresignedUrl(process.env.PROFILE_BUCKET, objectName, 'PUT')
-    ).url;
-    return { presignedUrl };
-  }
-
-  async putVideoPresignedUrl({ videoExtension, thumbnailExtension }) {
-    const videoId = new Types.ObjectId();
-    const [videoUrl, thumbnailUrl] = await Promise.all([
-      (
-        await createPresignedUrl(
-          process.env.INPUT_BUCKET,
-          `${videoId}.${videoExtension}`,
-          'PUT',
-        )
-      ).url,
-      (
-        await createPresignedUrl(
-          process.env.THUMBNAIL_BUCKET,
-          `${videoId}.${thumbnailExtension}`,
-          'PUT',
-        )
-      ).url,
-    ]);
-    return { videoId, videoUrl, thumbnailUrl };
-  }
-
-  async getImagePresignedUrl(id: string, { type, extension }) {
-    // TODO 업로드 된 이미지인지 확인
-    const bucketName = {
-      thumbnail: process.env.THUMBNAIL_BUCKET,
-      profile: process.env.PROFILE_BUCKET,
-    }[type];
-    const objectName = `${id}.${extension}`;
-    const presignedUrl = (
-      await createPresignedUrl(bucketName, objectName, 'GET')
-    ).url;
-    return { presignedUrl };
   }
 }
