+++ server/src/presigned-url/dto/advertisement-presigned-url-request.dto.ts
